move db dependencies
